handling responses from new_account

+++ clients/android/core/src/main/java/app/lockbook/core/core.kt
@@ -2,7 +2,7 @@ package app.lockbook.core
 
 external fun initLogger()
 external fun isDbPresent(path: String): Boolean
-external fun createAccount(path: String, username: String): Byte
+external fun createAccount(path: String, username: String): Int
